+++ client.go
@@ -145,6 +145,9 @@ func NewClientWithOptions(serverURL string, authToken string, options *Options) 
 		}
 		ilog.Infof("Using URL '%s'%s", serverURL, tokenStr)
 	}
+	if options.ApplicationName() == "" {
+		ilog.Warn("Application name is not set")
+	}
 	return client
 }